Address pull request comments
