+++ basyx.examples.timeseries/.gitignore
@@ -1,0 +1,1 @@
+/target/
