+++ basyx.examples.timeseries/Dockerfile
