+++ basyx.examples.timeseries/README.md
