+++ basyx.examples.timeseries/docker-compose.yaml
