+++ basyx.examples.timeseries/influxdb/.env
