+++ basyx.examples.timeseries/influxdb/.gitignore
