+++ basyx.examples.timeseries/influxdb/data_scripts/air-sensor-data.lp
