+++ basyx.examples.timeseries/influxdb/data_scripts/client.py
