+++ basyx.examples.timeseries/influxdb/entrypoint.sh
