+++ basyx.examples.timeseries/influxdb/scripts/influxdb-init.iql
