+++ basyx.examples.timeseries/pom.xml
