+++ basyx.examples.timeseries/src/main/resources/admin-shells/AirSensor.aasx
